backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
elasticsearch==8.10.1
pandas==2.1.2
pydantic-settings==2.1.0
python-multipart==0.0.6
xlsx2csv==0.8.1
passlib==1.7.4
python-jose==3.3.0
email-validator==2.0.0
cryptography==41.0.5
python-multipart==0.0.6
pyjwt==2.8.0